feat(frontend): migrate 3 pages to TanStack Query prefetch+useSuspenseQuery pattern #78
- repositories/[repositoryId]/page.tsx → Server prefetch + RepositoryDetailContent
- boards/[boardProjectId]/page.tsx → Server prefetch + BoardProjectDetailContent
- boards/[boardProjectId]/runs/page.tsx → Server prefetch + RunsListContent

Each page now uses:
- Server Component: prefetchQuery (no await) + HydrationBoundary
- Client Component: useSuspenseQuery for rendering

Error handling delegated to Error Boundary instead of notFound().
