+++ boardflow/src/app/(authenticated)/repositories/[repositoryId]/boards/[boardProjectId]/page.tsx
@@ -1,40 +1,10 @@
-import { Badge, Box, Heading, HStack, Table, Text, VStack } from '@chakra-ui/react';
-import Link from 'next/link';
-import { notFound } from 'next/navigation';
-import { Breadcrumb } from '@/components/ui/breadcrumb';
+import { Box } from '@chakra-ui/react';
+import { dehydrate, HydrationBoundary } from '@tanstack/react-query';
+import { Suspense } from 'react';
+import { BoardProjectDetailContent } from '@/components/board-project-detail/board-project-detail-content';
+import { $api } from '@/lib/api/react-query';
 import { createServerClient } from '@/lib/api/server';
-
-function statusColor(status: string): string {
-  switch (status) {
-    case 'completed':
-      return 'green';
-    case 'failed':
-      return 'red';
-    case 'timed_out':
-      return 'orange';
-    case 'created':
-    case 'uploading':
-    case 'importing':
-      return 'blue';
-    default:
-      return 'gray';
-  }
-}
-
-function checkBadge(status: string | null | undefined) {
-  if (!status)
-    return (
-      <Text color='gray.400' fontSize='sm'>
-        —
-      </Text>
-    );
-  const color = status === 'passed' ? 'green' : status === 'failed' ? 'red' : 'gray';
-  return (
-    <Badge colorPalette={color} size='sm'>
-      {status}
-    </Badge>
-  );
-}
+import { getQueryClient } from '@/lib/query-client';
 
 interface Props {
   params: Promise<{ repositoryId: string; boardProjectId: string }>;
@@ -42,199 +12,51 @@ interface Props {
 
 export default async function BoardProjectDetailPage({ params }: Props) {
   const { repositoryId, boardProjectId } = await params;
-  const client = await createServerClient();
+  const queryClient = getQueryClient();
+  const serverClient = await createServerClient();
 
-  const [{ data: project, error }, runsRes] = await Promise.all([
-    client.GET('/api/v1/board-projects/{board_project_id}', {
-      params: { path: { board_project_id: boardProjectId } },
-    }),
-    client.GET('/api/v1/board-projects/{board_project_id}/board-runs', {
+  const projectOptions = $api.queryOptions('get', '/api/v1/board-projects/{board_project_id}', {
+    params: { path: { board_project_id: boardProjectId } },
+  });
+
+  const runsOptions = $api.queryOptions(
+    'get',
+    '/api/v1/board-projects/{board_project_id}/board-runs',
+    {
       params: { path: { board_project_id: boardProjectId }, query: { limit: 5 } },
-    }),
-  ]);
+    },
+  );
 
-  if (error) {
-    notFound();
-  }
+  queryClient.prefetchQuery({
+    ...projectOptions,
+    queryFn: async () => {
+      const { data, error } = await serverClient.GET('/api/v1/board-projects/{board_project_id}', {
+        params: { path: { board_project_id: boardProjectId } },
+      });
+      if (error) throw new Error('Failed to fetch board project');
+      return data;
+    },
+  });
 
-  const recentRuns = runsRes.data?.items ?? [];
-  const runsLoadFailed = !!runsRes.error;
+  queryClient.prefetchQuery({
+    ...runsOptions,
+    queryFn: async () => {
+      const { data, error } = await serverClient.GET(
+        '/api/v1/board-projects/{board_project_id}/board-runs',
+        {
+          params: { path: { board_project_id: boardProjectId }, query: { limit: 5 } },
+        },
+      );
+      if (error) throw new Error('Failed to fetch board runs');
+      return data;
+    },
+  });
 
   return (
-    <Box>
-      <Breadcrumb
-        items={[
-          { label: 'Repositories', href: '/repositories' },
-          {
-            label: `${project.repository.owner}/${project.repository.name}`,
-            href: `/repositories/${repositoryId}`,
-          },
-          { label: project.display_name },
-        ]}
-      />
-      <VStack align='stretch' gap={6}>
-        <Box>
-          <HStack gap={2} mb={1}>
-            <Heading size='lg'>{project.display_name}</Heading>
-            <Badge colorPalette={project.state === 'completed' ? 'green' : 'gray'}>
-              {project.state}
-            </Badge>
-          </HStack>
-          <Text fontSize='sm' color='gray.600' fontFamily='mono'>
-            {project.project_path}
-          </Text>
-        </Box>
-
-        <Box borderWidth='1px' borderRadius='md' p={4} bg='white'>
-          <VStack align='stretch' gap={3}>
-            <HStack justify='space-between'>
-              <Text fontWeight='medium'>Repository</Text>
-              <Link href={`/repositories/${repositoryId}`}>
-                <Text color='blue.600' _hover={{ textDecoration: 'underline' }}>
-                  {project.repository.owner}/{project.repository.name}
-                </Text>
-              </Link>
-            </HStack>
-            <HStack justify='space-between'>
-              <Text fontWeight='medium'>Project Directory</Text>
-              <Text fontFamily='mono' fontSize='sm'>
-                {project.project_dir}
-              </Text>
-            </HStack>
-            {project.issue_url && (
-              <HStack justify='space-between'>
-                <Text fontWeight='medium'>Issue</Text>
-                <a href={project.issue_url} target='_blank' rel='noopener noreferrer'>
-                  <Text color='blue.600' _hover={{ textDecoration: 'underline' }}>
-                    #{project.issue_number}
-                  </Text>
-                </a>
-              </HStack>
-            )}
-            {project.latest_completed_run_id && (
-              <HStack justify='space-between'>
-                <Text fontWeight='medium'>Latest Completed Run</Text>
-                <Link
-                  href={`/repositories/${repositoryId}/boards/${boardProjectId}/runs/${project.latest_completed_run_id}`}
-                >
-                  <Text color='blue.600' _hover={{ textDecoration: 'underline' }}>
-                    {project.latest_completed_run_id}
-                  </Text>
-                </Link>
-              </HStack>
-            )}
-            <HStack justify='space-between'>
-              <Text fontWeight='medium'>Created</Text>
-              <Text fontSize='sm' color='gray.600'>
-                {new Date(project.created_at).toLocaleString()}
-              </Text>
-            </HStack>
-          </VStack>
-        </Box>
-
-        <Box>
-          <Heading size='md' mb={3}>
-            Recent Runs
-          </Heading>
-          {runsLoadFailed ? (
-            <Text color='red.500' fontSize='sm'>
-              Failed to load recent runs.
-            </Text>
-          ) : recentRuns.length === 0 ? (
-            <Text color='gray.500' fontSize='sm'>
-              No runs yet.
-            </Text>
-          ) : (
-            <Table.Root size='sm' variant='outline'>
-              <Table.Header>
-                <Table.Row>
-                  <Table.ColumnHeader>Status</Table.ColumnHeader>
-                  <Table.ColumnHeader>Commit</Table.ColumnHeader>
-                  <Table.ColumnHeader>Branch</Table.ColumnHeader>
-                  <Table.ColumnHeader>ERC</Table.ColumnHeader>
-                  <Table.ColumnHeader>DRC</Table.ColumnHeader>
-                  <Table.ColumnHeader>Created</Table.ColumnHeader>
-                </Table.Row>
-              </Table.Header>
-              <Table.Body>
-                {recentRuns.map((run) => (
-                  <Table.Row key={run.board_run_id}>
-                    <Table.Cell>
-                      <Link
-                        href={`/repositories/${repositoryId}/boards/${boardProjectId}/runs/${run.board_run_id}`}
-                      >
-                        <Badge colorPalette={statusColor(run.status)}>{run.status}</Badge>
-                      </Link>
-                    </Table.Cell>
-                    <Table.Cell>
-                      <Text fontFamily='mono' fontSize='sm'>
-                        {run.commit_sha.slice(0, 7)}
-                      </Text>
-                    </Table.Cell>
-                    <Table.Cell>
-                      <Text fontSize='sm'>{run.branch}</Text>
-                    </Table.Cell>
-                    <Table.Cell>
-                      <HStack gap={1}>
-                        {checkBadge(run.erc_status)}
-                        {run.erc_errors != null && run.erc_errors > 0 && (
-                          <Text fontSize='xs' color='red.500'>
-                            {run.erc_errors}E
-                          </Text>
-                        )}
-                        {run.erc_warnings != null && run.erc_warnings > 0 && (
-                          <Text fontSize='xs' color='orange.500'>
-                            {run.erc_warnings}W
-                          </Text>
-                        )}
-                      </HStack>
-                    </Table.Cell>
-                    <Table.Cell>
-                      <HStack gap={1}>
-                        {checkBadge(run.drc_status)}
-                        {run.drc_errors != null && run.drc_errors > 0 && (
-                          <Text fontSize='xs' color='red.500'>
-                            {run.drc_errors}E
-                          </Text>
-                        )}
-                        {run.drc_warnings != null && run.drc_warnings > 0 && (
-                          <Text fontSize='xs' color='orange.500'>
-                            {run.drc_warnings}W
-                          </Text>
-                        )}
-                      </HStack>
-                    </Table.Cell>
-                    <Table.Cell>
-                      <Text fontSize='sm' color='gray.600'>
-                        {new Date(run.created_at).toLocaleString()}
-                      </Text>
-                    </Table.Cell>
-                  </Table.Row>
-                ))}
-              </Table.Body>
-            </Table.Root>
-          )}
-        </Box>
-
-        <Box>
-          <Link href={`/repositories/${repositoryId}/boards/${boardProjectId}/runs`}>
-            <Box
-              display='inline-flex'
-              alignItems='center'
-              px={4}
-              py={2}
-              borderRadius='md'
-              bg='blue.600'
-              color='white'
-              fontWeight='medium'
-              fontSize='sm'
-              _hover={{ bg: 'blue.700' }}
-            >
-              View All Runs
-            </Box>
-          </Link>
-        </Box>
-      </VStack>
-    </Box>
+    <HydrationBoundary state={dehydrate(queryClient)}>
+      <Suspense fallback={<Box p={8}>Loading...</Box>}>
+        <BoardProjectDetailContent repositoryId={repositoryId} boardProjectId={boardProjectId} />
+      </Suspense>
+    </HydrationBoundary>
   );
 }
